Refactor: extract functions to fix check-file-line-limits CI (#1590)
Extract isStderrError to src/claude.stderr.lib.mjs and
checkForExistingComment/checkForNonBotComments to src/solve.pr-comments.lib.mjs
to bring src/claude.lib.mjs (1513→1478) and src/solve.auto-merge.lib.mjs
(1545→1363) under the 1500-line limit.
